[SPARK-42076][CONNECT][PYTHON] Factor data conversion `arrow -> rows` out to `conversion.py`
### What changes were proposed in this pull request?
Factor data conversion `arrow -> rows` out to `conversion.py`

### Why are the changes needed?
to make it more clearer, and also more consistent with PySpark

### Does this PR introduce _any_ user-facing change?
no, refactor-only

### How was this patch tested?
existing UT

Closes #39587 from zhengruifeng/connect_collect_refactor.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
